Added LoadActiveuser
Can be used in order to Cache the ActiveUser.Record object. This ensures that no performance is lost when using activeuser as a datasource on an actionpad.
This function should be called in ThisApplication.Setup. It will slow down the startup of LIME, but the users won't notice the performance dip when opening a card that uses activeuser as datasource the first time.